Merged PR 868: #5544 Fix sharing a track from a podcast
The problem was related with the name of the podcast which contained spaces in the name and a non encoded url was searched to be replaced.
F.e. there was no problem for the podcast MP3-Kilden because it has no space in the name.
Also, the problem was just for the podcasts. The private playlists and albums have the id in the url - which links to them.

Related work items: #5544